Extract google-drive IPC handlers to src/ipc/
main.ts has grown to 2,476 lines and 46 IPC handlers; the
google-drive cluster (5 handlers, ~233 lines of state + helpers
+ timer + OAuth flow) is the most self-contained chunk and can
move with zero behavior change.

Why a module-level singleton instead of pure dependency injection:
refreshGoogleSyncTimer is called from main.ts's applyConfigSideEffects
on every googleDriveEnabled/googleOAuth change, and maybeAutoSync is
called fire-and-forget from the post-save paths. Threading ctx through
both call sites would force every consumer to hold a ctx reference.
Stashing ctx on first register() keeps the import-site ergonomics
identical to the pre-refactor inline functions.

Why a separate src/ipc/types.ts and register.ts: two sister PRs are
extracting other clusters (meetings-mgmt, transcription) from the
same file in parallel. Each creates the same scaffolding so they can
land independently; the first to merge keeps it and the others rebase
to drop the duplicates and add their registration line.
